refactor(entranceForm): move details in Cave folder
